Move Python wrapping code for buffer to buffer/pymod/
