+++ cpp/modmesh/CMakeLists.txt
@@ -61,6 +61,9 @@ set(MODMESH_TERMINAL_FILES
     ${MODMESH_ROOT_PYMODHEADERS}
     ${MODMESH_ROOT_PYMODSOURCES}
     ${MODMESH_BUFFER_HEADERS}
+    ${MODMESH_BUFFER_SOURCES}
+    ${MODMESH_BUFFER_PYMODHEADERS}
+    ${MODMESH_BUFFER_PYMODSOURCES}
     ${MODMESH_MESH_HEADERS}
     ${MODMESH_MESH_SOURCES}
     ${MODMESH_DEVICE_HEADERS}
