+++ cpp/modmesh/buffer/pymod/buffer_pymod.hpp
@@ -40,11 +40,8 @@ namespace modmesh
 namespace python
 {
 
-void initialize_modmesh(pybind11::module & mod);
-void wrap_profile(pybind11::module & mod);
-void wrap_StaticGrid(pybind11::module & mod);
-void wrap_StaticMesh(pybind11::module & mod);
-void wrap_Toggle(pybind11::module & mod);
+void wrap_ConcreteBuffer(pybind11::module & mod);
+void wrap_SimpleArray(pybind11::module & mod);
 
 } /* end namespace python */
