+++ cpp/modmesh/buffer/pymod/wrap_ConcreteBuffer.cpp
@@ -26,7 +26,7 @@
  * POSSIBILITY OF SUCH DAMAGE.
  */
 
-#include <modmesh/python/wrapper/modmesh/modmesh.hpp> // Must be the first include.
+#include <modmesh/buffer/pymod/buffer_pymod.hpp> // Must be the first include.
 #include <modmesh/modmesh.hpp>
 
 namespace modmesh
